Update existing tests
